Changed challenge rating to a decimal, fixed 0 CR

+++ db/migrate/20180503163105_create_monsters.rb
@@ -5,7 +5,7 @@ class CreateMonsters < ActiveRecord::Migration[5.2]
       t.string :description
       t.string :size
       t.integer :hit_points
-      t.integer :challenge_rating
+      t.decimal :challenge_rating
       t.integer :strength
       t.integer :dexterity
       t.integer :constitution
